+++ CHANGELOG.md
@@ -11,6 +11,7 @@ ChangeLog
 - IMPROVEMENT: add "rnnoise" noise suppression node
 - IMPROVEMENT: add "amazon" speech to text node
 - IMPROVEMENT: add "compressor" audio processing node
+- IMPROVEMENT: add "awstranslate" text translation node
 
 1.4.5 (2025-08-07)
 ------------------